AcademicHistoryTableにEditAcademicHistoryModalを移植
